+++ src/main/java/de/zeanon/schemmanager/worldeditversion/commands/Folder.java
@@ -27,7 +27,7 @@ public class Folder {
 		new BukkitRunnable() {
 			@Override
 			public void run() {
-				int listmax = ConfigUtils.getInt("Listmax");
+				byte listmax = ConfigUtils.getByte("Listmax");
 				Path schemPath = WorldEditVersionSchemUtils.getSchemPath();
 				boolean spaceLists = ConfigUtils.getBoolean("Space Lists");
 
@@ -58,9 +58,9 @@ public class Folder {
 								MessageUtils.sendHoverMessage(ChatColor.AQUA + "=== ", ChatColor.AQUA + "" + (int) count + " Folder | Page 1/" + side, ChatColor.AQUA + " ===", ChatColor.GRAY + "Schematics", p);
 
 								if (count < listmax) {
-									listmax = (int) count;
+									listmax = (byte) count;
 								}
-								for (int i = 0; i < listmax; i++) {
+								for (byte i = 0; i < listmax; i++) {
 									if (sendListLineFailed(p, schemPath, listPath, files[i], i, deepSearch)) {
 										return;
 									}
@@ -105,9 +105,9 @@ public class Folder {
 
 									int id = (side_number - 1) * listmax;
 									if (count < listmax * side_number) {
-										listmax = (int) count - (listmax * (side_number - 1));
+										listmax = (byte) ((int) count - (listmax * (side_number - 1)));
 									}
-									for (int i = 0; i < listmax; i++) {
+									for (byte i = 0; i < listmax; i++) {
 										if (sendListLineFailed(p, schemPath, listPath, files[id], id, deepSearch)) {
 											return;
 										}
@@ -154,9 +154,9 @@ public class Folder {
 									MessageUtils.sendHoverMessage(ChatColor.AQUA + "=== ", ChatColor.AQUA + "" + (int) count + " Folder | Page 1/" + side, ChatColor.AQUA + " ===", ChatColor.GRAY + "Schematics/" + args[2], p);
 
 									if (count < listmax) {
-										listmax = (int) count;
+										listmax = (byte) count;
 									}
-									for (int i = 0; i < listmax; i++) {
+									for (byte i = 0; i < listmax; i++) {
 										if (sendListLineFailed(p, schemPath, listPath, files[i], i, deepSearch)) {
 											return;
 										}
@@ -200,9 +200,9 @@ public class Folder {
 
 								int id = (side_number - 1) * listmax;
 								if (count < listmax * side_number) {
-									listmax = (int) count - (listmax * (side_number - 1));
+									listmax = (byte) ((int) count - (listmax * (side_number - 1)));
 								}
-								for (int i = 0; i < listmax; i++) {
+								for (byte i = 0; i < listmax; i++) {
 									if (sendListLineFailed(p, schemPath, listPath, files[id], id, deepSearch)) {
 										return;
 									}
