+++ src/main/java/de/zeanon/schemmanager/worldeditversion/commands/List.java
@@ -28,7 +28,7 @@ public class List {
 		new BukkitRunnable() {
 			@Override
 			public void run() {
-				int listmax = ConfigUtils.getInt("Listmax");
+				byte listmax = ConfigUtils.getByte("Listmax");
 				Path schemPath = WorldEditVersionSchemUtils.getSchemPath();
 				boolean spaceLists = ConfigUtils.getBoolean("Space Lists");
 				String[] extensions = ConfigUtils.getStringList("File Extensions").toArray(new String[0]);
@@ -60,9 +60,9 @@ public class List {
 								MessageUtils.sendHoverMessage(ChatColor.AQUA + "=== ", ChatColor.AQUA + "" + (int) count + " Schematics | Page 1/" + side, ChatColor.AQUA + " ===", ChatColor.GRAY + "Schematics", p);
 
 								if (count < listmax) {
-									listmax = (int) count;
+									listmax = (byte) count;
 								}
-								for (int i = 0; i < listmax; i++) {
+								for (byte i = 0; i < listmax; i++) {
 									if (sendListLineFailed(p, schemPath, listPath, files[i], i, deepSearch)) {
 										return;
 									}
@@ -107,9 +107,9 @@ public class List {
 
 									int id = (side_number - 1) * listmax;
 									if (count < listmax * side_number) {
-										listmax = (int) count - (listmax * (side_number - 1));
+										listmax = (byte) ((int) count - (listmax * (side_number - 1)));
 									}
-									for (int i = 0; i < listmax; i++) {
+									for (byte i = 0; i < listmax; i++) {
 										if (sendListLineFailed(p, schemPath, listPath, files[id], id, deepSearch)) {
 											return;
 										}
@@ -156,9 +156,9 @@ public class List {
 									MessageUtils.sendHoverMessage(ChatColor.AQUA + "=== ", ChatColor.AQUA + "" + (int) count + " Schematics | Page 1/" + side, ChatColor.AQUA + " ===", ChatColor.GRAY + "Schematics/" + args[2], p);
 
 									if (count < listmax) {
-										listmax = (int) count;
+										listmax = (byte) count;
 									}
-									for (int i = 0; i < listmax; i++) {
+									for (byte i = 0; i < listmax; i++) {
 										if (sendListLineFailed(p, schemPath, listPath, files[i], i, deepSearch)) {
 											return;
 										}
@@ -203,9 +203,9 @@ public class List {
 
 								int id = (side_number - 1) * listmax;
 								if (count < listmax * side_number) {
-									listmax = (int) count - (listmax * (side_number - 1));
+									listmax = (byte) ((int) count - (listmax * (side_number - 1)));
 								}
-								for (int i = 0; i < listmax; i++) {
+								for (byte i = 0; i < listmax; i++) {
 									if (sendListLineFailed(p, schemPath, listPath, files[id], id, deepSearch)) {
 										return;
 									}
